Allow rpm tools on RHEL < 8 to parse the spec file
We use some automated tools to parse the spec file, which do not
recognize the newer "Recommends" and "Suggests" tags because they
run on RHEL 7. To allow these tools to work while keeping the
functionality in the spec, let's enclose the tags in an %if clause.

This is similar to https://review.rdoproject.org/r/32035

Change-Id: I0ce9343f9ab99515f3b00935a02ceab11bb4c1e0